requirements.txt
flask
Flask-Limiter
flask_sqlalchemy
pre-commit
pydantic
pydantic[email]
pydantic-sqlalchemy
pytest
python-dotenv
requests
